refactor: extract members into own classes
